requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2
python-dotenv
